% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bVAR.R
\name{bVAR}
\alias{bVAR}
\title{R6 class of bayesian VAR}
\description{
bVAR: bayesian VAR model
}
\details{
The \code{bVAR} class implements bayesian VAR estimation an identification.
\itemize{
\item estimation: conjugate prior or gibbs sampling
\item identification: sign restrictions
\item inference: bayesian HPD
}
}
\seealso{
\code{\link[=VAR$IRF.compute]{VAR$IRF.compute()}}
}
\section{Super class}{
\code{VARS::basic VAR model} -> \code{Bayes VAR}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prior.type}}{type of prior, can be "info", "flat" or "user"}

\item{\code{alpha.prior}}{store prior of VAR coefficients}

\item{\code{Sigma.prior}}{store prior of VAR Cov matrix}

\item{\code{alpha.post}}{store post of VAR coefficients}

\item{\code{Sigma.post}}{store post of VAR Cov matrix}

\item{\code{B.NSR}}{save impact matrices of SVAR with narrative sign restrictions}

\item{\code{Sigma.NSR}}{save Cov matrices of VAR with narrative sign restrictions}

\item{\code{beta.NSR}}{save coefficients of VAR with narrative sign restrictions}

\item{\code{get.result.tmp}}{indicator of whether SR and EBR are imposed}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Bayes VAR-new}{\code{bVAR$new()}}
\item \href{#method-Bayes VAR-est}{\code{bVAR$est()}}
\item \href{#method-Bayes VAR-identify.seq}{\code{bVAR$identify.seq()}}
\item \href{#method-Bayes VAR-tool}{\code{bVAR$tool()}}
\item \href{#method-Bayes VAR-IRF.compute.NSR}{\code{bVAR$IRF.compute.NSR()}}
\item \href{#method-Bayes VAR-clone}{\code{bVAR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="FEVD.compute"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-FEVD.compute'><code>VARS::basic VAR model$FEVD.compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="FEVD.compute.batch"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-FEVD.compute.batch'><code>VARS::basic VAR model$FEVD.compute.batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="HDC.compute"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-HDC.compute'><code>VARS::basic VAR model$HDC.compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="HDC.compute.batch"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-HDC.compute.batch'><code>VARS::basic VAR model$HDC.compute.batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="IRF.compute"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-IRF.compute'><code>VARS::basic VAR model$IRF.compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="IRF.compute.batch"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-IRF.compute.batch'><code>VARS::basic VAR model$IRF.compute.batch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bayes VAR-new"></a>}}
\if{latex}{\out{\hypertarget{method-Bayes VAR-new}{}}}
\subsection{Method \code{new()}}{
constructor of \code{bVAR}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bVAR$new(data = NA, p.lag = NA, prior.type = "info", priors = NA, cons = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{time series data}

\item{\code{p.lag}}{number of lags in model}

\item{\code{prior.type}}{if \code{priors} not provided, specify prior type, "flat"(Normal) or "info"(OLS result)}

\item{\code{priors}}{user provided prior, \code{prior.type} will be set to "user"}

\item{\code{cons}}{whether include constant in model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an R6 class of \code{bVAR}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bayes VAR-est"></a>}}
\if{latex}{\out{\hypertarget{method-Bayes VAR-est}{}}}
\subsection{Method \code{est()}}{
estimation of bVAR model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bVAR$est(
  Y = self$Y,
  X = self$X,
  method = "gibbs",
  burn.in = 100,
  draw = 1000,
  thin = 1,
  post.save = NA,
  post.method = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{\code{self$Y}}

\item{\code{X}}{\code{self$X}}

\item{\code{method}}{"conjugate" or "gibbs"}

\item{\code{burn.in}}{required when \code{method=="gibbs"}, dropping samples before save}

\item{\code{draw}}{required when \code{method=="gibbs"}, number of saved draws}

\item{\code{thin}}{required when \code{method=="gibbs"}}

\item{\code{post.save}}{required when \code{method=="gibbs"}, save post distribution parameters or not}

\item{\code{post.method}}{required when \code{method=="gibbs"}, how to save post distribution parameters (0-last, 1-mean)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, results saved in \verb{self$*.post}, \code{self$beta} and \code{self$Sigma}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bayes VAR-identify.seq"></a>}}
\if{latex}{\out{\hypertarget{method-Bayes VAR-identify.seq}{}}}
\subsection{Method \code{identify.seq()}}{
identification of bVAR model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bVAR$identify.seq(SR = NA, EBR = NA, NSR = NA, draw = 5000, M = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SR}}{sign restrictions}

\item{\code{EBR}}{elasticity bound restrictions}

\item{\code{NSR}}{narrative sign restrictions}

\item{\code{draw}}{number of required number of saved samples}

\item{\code{M}}{number of samples to draw when simulating importance weight}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, results saved in \verb{self$*.set} and \verb{self$*.NSR}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bayes VAR-tool"></a>}}
\if{latex}{\out{\hypertarget{method-Bayes VAR-tool}{}}}
\subsection{Method \code{tool()}}{
compute VAR tools, currently support IRF, FEVD and HDC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bVAR$tool(
  tool = NA_character_,
  hor = NA_integer_,
  prob = NA_real_,
  start = NA_integer_,
  end = NA_integer_,
  which.var = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tool}}{name of VAR tools to be computed}

\item{\code{hor}}{horizons}

\item{\code{prob}}{probability of CI}

\item{\code{start}}{start of interested period, only required when `tool=="HDC"'}

\item{\code{end}}{end of interested period, only required when `tool=="HDC"'}

\item{\code{which.var}}{the variable name to do HDC, only required when `tool=="HDC"'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
related VAR tools, returned as a list with it's average, median and upper lower bound values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bayes VAR-IRF.compute.NSR"></a>}}
\if{latex}{\out{\hypertarget{method-Bayes VAR-IRF.compute.NSR}{}}}
\subsection{Method \code{IRF.compute.NSR()}}{
compute IRFs of NSR parameter sets, used for paper replic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bVAR$IRF.compute.NSR(hor = NA_integer_, prob = NA_real_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hor}}{horizon}

\item{\code{prob}}{probability of CI}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list, containing median, average, upper bound and lower bound of the CI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bayes VAR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Bayes VAR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bVAR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
