\name{coef}

\alias{coef}

\alias{coefficients}

\alias{coef.varest}

\encoding{latin1}

\title{Coefficient method for objects of class varest}

\description{
  Returns the coefficients of a VAR(p)-model for objects generated by
  \command{VAR()}. Thereby the coef-method is applied to the summary of
  the list element \code{varresult}, which is itself a list of
  \code{summary.lm}-objects. 
}

\usage{
\method{coef}{varest}(object, ...)
}

\arguments{
  \item{object}{An object of class \sQuote{\code{varest}}; generated by
    \command{VAR()}} 
  \item{\dots}{Currently not used.}
}

\author{Bernhard Pfaff}

\references{
  Hamilton, J. (1994), \emph{Time Series Analysis}, Princeton
  University Press, Princeton.

  Lütkepohl, H. (2006), \emph{New Introduction to Multiple Time Series
    Analysis}, Springer, New York.  
}

\seealso{
  \code{\link{VAR}}
}

\examples{
data(Canada)
var.2c <- VAR(Canada, p = 2, type = "const")
coef(var.2c)
}

\keyword{regression}
