% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cVAR.R
\name{cVAR}
\alias{cVAR}
\title{R6 class of classic VAR}
\description{
cVAR: classic VAR model
}
\details{
The \code{cVAR} class implements classic VAR estimation an identification.
\itemize{
\item estimation: OLS
\item identification: recursive and IV
\item inference: bootstrap
}
}
\section{Super class}{
\code{VARS::basic VAR model} -> \code{Classic VAR}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rank}}{store rank of OLS \verb{XX'} matrix}

\item{\code{est.method}}{store name of estimation method, should be "OLS"}

\item{\code{identify.method}}{store name of identify method, should be "recursive" or "IV"}

\item{\code{IV}}{store IV data series, should be named matrix}

\item{\code{IV.provide}}{bool, indicate whether IV is provided}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Classic VAR-new}{\code{cVAR$new()}}
\item \href{#method-Classic VAR-est}{\code{cVAR$est()}}
\item \href{#method-Classic VAR-identify}{\code{cVAR$identify()}}
\item \href{#method-Classic VAR-tool}{\code{cVAR$tool()}}
\item \href{#method-Classic VAR-clone}{\code{cVAR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="FEVD.compute"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-FEVD.compute'><code>VARS::basic VAR model$FEVD.compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="FEVD.compute.batch"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-FEVD.compute.batch'><code>VARS::basic VAR model$FEVD.compute.batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="HDC.compute"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-HDC.compute'><code>VARS::basic VAR model$HDC.compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="HDC.compute.batch"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-HDC.compute.batch'><code>VARS::basic VAR model$HDC.compute.batch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="IRF.compute"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-IRF.compute'><code>VARS::basic VAR model$IRF.compute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="VARS" data-topic="basic VAR model" data-id="IRF.compute.batch"><a href='../../VARS/html/basic VAR model.html#method-basic VAR model-IRF.compute.batch'><code>VARS::basic VAR model$IRF.compute.batch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic VAR-new"></a>}}
\if{latex}{\out{\hypertarget{method-Classic VAR-new}{}}}
\subsection{Method \code{new()}}{
\code{cVAR} class constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cVAR$new(data = NA, p.lag = NA, data.IV = NA, cons = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{time series data frame or names matrix}

\item{\code{p.lag}}{int, number of lags in model}

\item{\code{data.IV}}{named matrix, the IV series; the IV name should \strong{be set as the same as the related variable name}}

\item{\code{cons}}{bool, contain constant or not in model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an R6 class, \code{cVAR}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic VAR-est"></a>}}
\if{latex}{\out{\hypertarget{method-Classic VAR-est}{}}}
\subsection{Method \code{est()}}{
estimate reduced form VAR using OLS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cVAR$est(method = "OLS")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{name of estimation method, currently only support "OLS"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, estimation result saved in \code{self$beta}, \code{self$U} and \code{self$Sigma}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic VAR-identify"></a>}}
\if{latex}{\out{\hypertarget{method-Classic VAR-identify}{}}}
\subsection{Method \code{identify()}}{
idenitfy classic VAR model using "recursive" or "IV" method, then do bootstrap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cVAR$identify(
  method = NA_character_,
  boot.method = NA_character_,
  repeats = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{identification method, should be "recursive" or "IV"}

\item{\code{boot.method}}{bootstrap method, should be "wild" or "block"}

\item{\code{repeats}}{bootstrap repeat times}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, results saved in \code{self$B}, \code{self$eps} and \verb{self$*.set}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic VAR-tool"></a>}}
\if{latex}{\out{\hypertarget{method-Classic VAR-tool}{}}}
\subsection{Method \code{tool()}}{
compute VAR tools, currently support IRF, FEVD and HDC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cVAR$tool(
  tool = NA_character_,
  hor = NA_integer_,
  prob = NA_real_,
  start = NA_integer_,
  end = NA_integer_,
  which.var = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tool}}{name of VAR tools to be computed}

\item{\code{hor}}{horizons}

\item{\code{prob}}{probability of CI}

\item{\code{start}}{start of interested period, only required when `tool=="HDC"'}

\item{\code{end}}{end of interested period, only required when `tool=="HDC"'}

\item{\code{which.var}}{the variable name to do HDC, only required when `tool=="HDC"'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
related VAR tools, returned as a list with it's average, median and upper lower bound values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic VAR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Classic VAR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cVAR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
