\name{fanchart}

\alias{fanchart}

\title{Fanchart plot for objects of class varprd}

\description{
  Time Series plots of VAR forecasts with differently shaded confidence
  regions (fanchart) for each endogenous variable. 
}

\usage{
fanchart(x, colors = NULL, cis = NULL, names = NULL, main = NULL, ylab =
NULL, xlab = NULL, col.y = NULL, nc, plot.type = c("multiple",
"single"), mar = par("mar"), oma = par("oma"), ... )
}

\arguments{
  \item{x}{An object of class \sQuote{\code{varprd}}; generated by
    \command{predict()}.}
  \item{colors}{Character vector of colors to be used for shading. If
    unset, a gray color scheme is used.}
  \item{cis}{A numeric vector of confidence intervals. If unset the
    sequence from 0.1 to 0.9 is used (step size 0.1).}
  \item{names}{Character vector, names of variables for fancharts. If
    unset, all variables are plotted.}
  \item{main}{Character vector, title for fanchart plots.}
  \item{ylab}{Character, ylab for fanchart.}
  \item{xlab}{Character, xlab for fanchart.}
  \item{col.y}{Character, color for plotted time series.}
  \item{plot.type}{Character, if multiple all fancharts appear in one
    device.}
  \item{nc}{Integer, number of columns if plot.type is multiple.}
  \item{mar}{Vector, setting of margins.}
  \item{oma}{Vector, setting of outer margins.}
  \item{...}{Dot argument, passed to plot.ts.}
}

\author{Bernhard Pfaff}

\references{
  Britton, E., P.G. Fisher and J.D. Whitley (1998), Inflation Report
  projections: understanding the fan chart, \emph{Bank of England
    Quarterly Bulletin}, February, Bank of England, pages 30-37.
}

\seealso{
  \code{\link{VAR}}, \code{\link{predict}}, \code{\link{plot}},
  \code{\link{par}}  
}

\examples{
\dontrun{
data(Canada)
var.2c <- VAR(Canada, p = 2, type = "const")
var.2c.prd <- predict(var.2c, n.ahead = 8, ci = 0.95)
fanchart(var.2c.prd)
}
}

\keyword{regression}
\concept{Fanchart}
