\name{fitted}

\alias{fitted}

\alias{fitted.values}

\alias{fitted.varest}

\alias{fitted.vec2var}

\encoding{latin1}

\title{Fit method for objects of class varest or vec2var}

\description{
  Returns the fitted values of a VAR(p)-model for objects generated by
  \command{VAR()} or \command{vec2var()}. For objects of class varest
  the fitted.values-method is applied to the list element
  \code{varresult}, which is itself a list of \code{lm}-objects. 
}

\usage{
\method{fitted}{varest}(object, ...)
\method{fitted}{vec2var}(object, ...)
}

\arguments{
  \item{object}{An object of class \sQuote{\code{varest}}; generated by
    \command{VAR()}, or an object of class \sQuote{\code{vec2var}};
    generated by \command{vec2var()}.} 
  \item{\dots}{Currently not used.}
}

\author{Bernhard Pfaff}

\references{
  Hamilton, J. (1994), \emph{Time Series Analysis}, Princeton
  University Press, Princeton.

  Lütkepohl, H. (2006), \emph{New Introduction to Multiple Time Series
    Analysis}, Springer, New York.  
}

\seealso{
  \code{\link{VAR}}, \code{\link{vec2var}}
}

\examples{
data(Canada)
var.2c <- VAR(Canada, p = 2, type = "const")
fitted(var.2c)
}

\keyword{regression}
