% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GK2015}
\alias{GK2015}
\title{Gertler and Karadi (2015) dataset}
\format{
\subsection{GK2015}{

The data used in Gertler and Karadi (2015) paper \emph{Monetary Policy Surprises, Credit Costs, and Economic Activity}, monthly data, 396 obs of 6 variables.
\describe{
\item{date}{month indicator, 1979m7 - 2012m6}
\item{logcpi}{log of USA CPI}
\item{logip}{log of USA industrial production}
\item{gs1}{1 year USA government bond rate, the policy variable}
\item{ebp}{excess bond premium}
\item{ff4_tc}{three month ahead monthly fed funds futures, the IV}
}
}
}
\source{
http://doi.org/10.3886/E114082V1
}
\usage{
data(GK2015)
}
\description{
The data used in Gertler and Karadi (2015) paper \emph{Monetary Policy Surprises, Credit Costs, and Economic Activity}.
}
\keyword{datasets}
