% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IRF.plot}
\alias{IRF.plot}
\title{Plot IRF}
\usage{
IRF.plot(obj, IRFs, hor, shk.names, var.names)
}
\arguments{
\item{obj}{a cVAR or bVAR object, provide full variable names}

\item{IRFs}{IRF computed from cVAR or bVAR}

\item{hor}{horizons you want to plot}

\item{shk.names}{which shocks you want to include}

\item{var.names}{which variables you want to include}
}
\value{
a figure with sub figues, each containing an IRF of 1 variable, 1 shock
}
\description{
Plot IRF
}
\seealso{
\code{\link[=bVAR$tool]{bVAR$tool()}}, \code{\link[=cVAR$tool]{cVAR$tool()}}
}
