% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kilian2012}
\alias{Kilian2012}
\title{Kilian and Murphy (2012) dataset}
\format{
\subsection{Kilian2012}{

The data used in Kilian and Murphy (2012) paper \emph{Why agnostic sign restrictions are not enough: understanding the dynamics of oil market var models}, monthly data, 540 obs of 4 variables
\describe{
\item{Date}{month indicator, 1971m1 - 2015m12}
\item{Oil.Production.Growth}{global oil production growth}
\item{Economic.Activity.Index}{global economic activity index}
\item{Real.Oil.Price}{global real oil price}
}
}
}
\source{
https://doi.org/10.1111/j.1542-4774.2012.01080.x
}
\usage{
data(Kilian2012)
}
\description{
The data used in Kilian and Murphy (2012) paper \emph{Why agnostic sign restrictions are not enough: understanding the dynamics of oil market var models}
}
\keyword{datasets}
