% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat2cube}
\alias{mat2cube}
\title{Transforming IRF matrix to 3D array}
\usage{
mat2cube(mat)
}
\arguments{
\item{mat}{The IRF matrix, row variable shock combination, col horizon}
}
\value{
a 3D array, with each slice the IRF matrix of that horizon
}
\description{
Transforming IRF matrix to 3D array
}
\seealso{
\code{\link[=IRF.compute.batch]{IRF.compute.batch()}}
}
