% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions.R
\name{NoutofT}
\alias{NoutofT}
\title{This function selects randomly N elements out of a set of T elements}
\usage{
NoutofT(N, T)
}
\arguments{
\item{N}{: the number to be selected}

\item{T}{: the total number of elements}
}
\description{
This function selects randomly N elements out of a set of T elements
}
