\name{plot.varest}

\alias{plot.varest}

\alias{plot.vec2var}

\encoding{latin1}

\title{Plot method for objects of class varest or vec2var}

\description{
  Plots per endogenous variable a diagram of fit, the residuals and
  their ACFs und PACFs for an object with class attribute
  \sQuote{\code{varest}}; generated by \command{VAR()}, or with class attribute
  \sQuote{\code{vec2var}}; generated by \command{vec2var()}.
}

\usage{
\method{plot}{varest}(x, ...)
\method{plot}{vec2var}(x, ...)
}

\arguments{
  \item{x}{An object of class \sQuote{\code{varest}}; generated by
    \command{VAR()}, or an object of class \sQuote{\code{vec2var}}; generated by
    \command{vec2var()}.}
  \item{\dots}{Currently not used.}
}

\author{Bernhard Pfaff}

\references{
  Hamilton, J. (1994), \emph{Time Series Analysis}, Princeton
  University Press, Princeton.

  Lütkepohl, H. (2006), \emph{New Introduction to Multiple Time Series
    Analysis}, Springer, New York.  
}

\seealso{
  \code{\link{VAR}}
}

\examples{
\dontrun{
data(Canada)
var.2c <- VAR(Canada, p = 2, type = "const")
plot(var.2c)
}
}

\keyword{regression}
