% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions.R
\name{rnormSIGMA}
\alias{rnormSIGMA}
\title{This function will generate iid multivariate normal random time series.}
\usage{
rnormSIGMA(T, sigma)
}
\arguments{
\item{T}{: length of the generated time series}

\item{sigma}{: (n x n) covariance matrix of the normal series}
}
\value{
\preformatted{ : T x n matrix of iid normal time series
}
}
\description{
This function will generate iid multivariate normal random time series.
}
