% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SW2001}
\alias{SW2001}
\title{Stock and Watson (2001) dataset}
\format{
\subsection{SW2001}{

The data used in Stock and Watson (2001) paper \emph{Vector Autoregressions}, quarterly data, 164 obs. of 4 variables.
\describe{
\item{date}{quarter indicator, 1960q1 - 2000q4}
\item{unemp}{USA unemployment rate}
\item{infl}{USA inflation rate}
\item{ffr}{USA federal funds rate}
}
}
}
\source{
https://www.aeaweb.org/articles?id=10.1257/jep.15.4.101
}
\usage{
data(SW2001)
}
\description{
The data used in Stock and Watson (2001) paper \emph{Vector Autoregressions}.
}
\keyword{datasets}
