% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Uhlig2005}
\alias{Uhlig2005}
\title{Uhlig (2005) dataset}
\format{
\subsection{Uhlig2005}{

The data used in Uhlig (2005) paper \emph{What are the effects of monetary policy on output? Results from an agnostic identification procedure}, monthly data, 515 obs of 7 variables
\describe{
\item{Date}{month indicator, 1965m1 - 2007m11}
\item{GDPC1}{real GDP of USA}
\item{GDPDEF}{GDP deflator}
\item{CPRINDEX}{CPI}
\item{TOTRESNS}{total reserves}
\item{BOGNONBR}{non borrowed reserves}
\item{FEDFUNDS}{federal funds rate}
}
}
}
\source{
https://doi.org/10.1016/j.jmoneco.2004.05.007
}
\usage{
data(Uhlig2005)
}
\description{
The data used in Uhlig (2005) paper \emph{What are the effects of monetary policy on output? Results from an agnostic identification procedure}
}
\keyword{datasets}
