\name{vars-deprecated}
\alias{vars-deprecated}
\alias{A}
\alias{B}
\alias{arch}
\alias{normality}
\alias{serial}
\docType{data}
\title{Deprecated Functions in package vars}
\description{
  These functions are provided for compatibility with older versions of package vars only, and may be defunct as soon as the next release.
}
\details{
  'A' is a deprecated synonym for 'Acoef'.\cr
  'arch' is a deprecated synonym for 'arch.test'.\cr
  'B' is a deprecated synonym for 'Bcoef'.\cr
  'normality' is a deprecated synonym for 'normality.test'.\cr
  'serial' is a deprecated synonym for 'serial.test'.
}

\seealso{
    \code{\link{Acoef}}, \code{\link{arch.test}}, \code{\link{Bcoef}},
    \code{\link{normality.test}}, \code{\link{serial.test}}  
}

\keyword{regression}
