% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodel.R
\name{VAR}
\alias{VAR}
\title{VAR: basic VAR model}
\description{
The \code{VAR} class implements basic data manipulation, info extraction and VAR tools computation, which are inherited by other specific VAR classes.
}
\section{Methods}{

\describe{
\item{\code{new(name, value)}}{Creates a new instance of the class.}
\item{\code{greet()}}{Prints a greeting message.}
\item{\code{add(x, y)}}{Adds two numbers and returns the result.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{T}}{int, total time periods of data}

\item{\code{cons}}{bool, indicating whether include constant in model}

\item{\code{T.est}}{int, total time periods used for estimation}

\item{\code{p.lag}}{int, lags of the VAR model}

\item{\code{hor}}{int, number of horizons interested in inference}

\item{\code{data}}{\strong{named} matrix of dataframe, each column indicating one time series, the first column should be time stramp}

\item{\code{var.names}}{character vector, variable names}

\item{\code{time}}{character of Date vector, time stramp}

\item{\code{Y}}{numeric matrix, dependent variables}

\item{\code{Y.start}}{numeric matrix, first \code{p.lag} periods of data}

\item{\code{X}}{numeric matrix, dependent variables}

\item{\code{beta}}{numeric matrix, VAR reduced form coefficients}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-basic VAR model-new}{\code{VAR$new()}}
\item \href{#method-basic VAR model-IRF.compute}{\code{VAR$IRF.compute()}}
\item \href{#method-basic VAR model-FEVD.compute}{\code{VAR$FEVD.compute()}}
\item \href{#method-basic VAR model-HDC.compute}{\code{VAR$HDC.compute()}}
\item \href{#method-basic VAR model-IRF.compute.batch}{\code{VAR$IRF.compute.batch()}}
\item \href{#method-basic VAR model-FEVD.compute.batch}{\code{VAR$FEVD.compute.batch()}}
\item \href{#method-basic VAR model-HDC.compute.batch}{\code{VAR$HDC.compute.batch()}}
\item \href{#method-basic VAR model-clone}{\code{VAR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-new"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$new(data = NA, p.lag = NA, cons = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-IRF.compute"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-IRF.compute}{}}}
\subsection{Method \code{IRF.compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$IRF.compute(hor = NA_integer_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-FEVD.compute"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-FEVD.compute}{}}}
\subsection{Method \code{FEVD.compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$FEVD.compute(hor = self$hor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-HDC.compute"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-HDC.compute}{}}}
\subsection{Method \code{HDC.compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$HDC.compute(start = NA, end = NA, which.var = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-IRF.compute.batch"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-IRF.compute.batch}{}}}
\subsection{Method \code{IRF.compute.batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$IRF.compute.batch(hor = NA_integer_, prob = NA_real_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-FEVD.compute.batch"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-FEVD.compute.batch}{}}}
\subsection{Method \code{FEVD.compute.batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$FEVD.compute.batch(hor = NA_integer_, prob = NA_real_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-HDC.compute.batch"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-HDC.compute.batch}{}}}
\subsection{Method \code{HDC.compute.batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$HDC.compute.batch(start = NA, end = NA, which.var = NA_character_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basic VAR model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-basic VAR model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VAR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
